Solve IntegrationError
